core: make sockaddr_union* and socket_info* arguments, vars and members const
Make the said pointers const in most places where appropriate. This basically
ensures that code in modules cannot mess with them intentionally or by a
mistake.

It should also have some positive effect on the performance of the code, as
with those being const * the optimizer would have more room to speculate
about what code might and might not do.

There might be also some improvement of static analysis front.

Most of the change is pretty mechanical, the only functional change is
the allocation of the socket_info.last_real_port, since those are expected
to be updated by the underlying protos/modules. Work around that by
making a separate smaller struct last_real_ports and co-allocate it along
with the socket_info main backing storage, exposing the non-const pointer
to that instead.

Move the prev / next into the semi-private part of the struct.

Rework net/api_proto_net.h to not require casting handler functions
to a generic type.

Next step would be make str members of socket_info into str_const.

Tested with voiptests.